Add Error Code to String Lookup Utility
You'll have to just trust me that this works (it does). It unfortunately means that the error enumeration has to be declared in a rather weird way in order to be able to generate the string equivalents and lookup tables automatically. But what's the preprocessor for if not crazy stuff like this!

Also add a predefined region for custom error codes in case people wish to add their own. The standard error codes should never be modified as they are shared across multiple projects.