Util/lowlevel.h: Include ACLE intrinsics header
`__arm_mcr` and `__arm_mrc` are declared in `arm_acle.h`, which was not
explicitly included in `lowlevel.h`.  This was causing compilation
failures where symbols `__arm_mcr` and `__arm_mrc` could not be found.

Error observed on the lab machines, which are running a slightly older
version of the Arm compiler.  I suspect on newer compilers the header
is included transitively somehow, hence this going unnoticed before now.