Move SVC Handler to startup_arm
This actually has nothing to do with the normal IRQs, so move it out to live with the vector table and associated code.

Furthermore, correct the behaviour of the handler to cope better when semi-hosting is disabled and it finds itself trying to filter out debugger calls that are thrown its way.