Save/Restore SPSR State Before/After IRQ Handler Calls
Because the user IRQ handlers are tagged with __irq, they include epilogue for restoring the SPSR register to the CPSR on return. However because we are already in the IRQ mode at this point, that save/restore clobbers the SPSR value that we need restoring when returning from __irq_isr, causing weird behaviour. It will also cause problems in main irq handler itself as the CPSR register gets clobbered on return too.

To combat this, we backup the SPSR register at the start of the __irq_isr handler, and restore at the end just before returning. This will prevent issues with the returned state.