Only Change CoProc Regs if Required
When calling back to launch a new image, something has happened which prevents the __SET_SYSREG function from executing (becomes undefined). As such, only try to change the register if not already set correctly.
This means the __SET_SYSREG will be performed only by the first application to run (as all share the same settings, this is fine).

While at it, tidy up the accesses to CPACR to use defined values rather than nondescript constants.