New debugload.ds version to address error 'waitForTargetToStop' connecting to Debugger
Some DE1-SoC boards are experiencing an error whereby upon connecting to the debugger they are stuck reporting 'waitForTargetToStop'.

This new code attempts to reset the hardware to a know state three times during the debug connection as it appears some boards were not responding to the previous single reset.