Split Semihosting Code from Startup
The failback semihosting was a simple SVC IRQ that did nothing to allow startup when the debugger is not connected. Helpful, but not very featureful.

Now we have an entire failback semihosting library that allows e.g. redirecting semihosting printf calls to a UART instance if the debugger is not connected.

Furthermore, fully elaborate how to implement software interrupts. The new split now also allows proper use of user SWIs at the same time as or independently of the semihosting failback.

Since changing to armclang from armcc, the calling convention of SWIs has changed meaning a simple tag on a function is no longer enough to ensure it is called correctly. Instead we have to create our own calling stub. To simplify this, a macro has been added which is used to create the body of the stub, and documentation on how it is used is added.