Correct Event_create Table Index Calculation
When extending the table, the new index was being set incorrectly to an index beyond the range of the table.

Also we have a defined macro for no-op interval but it was not being checked against (technically it still worked because the macro is defined as 0, but let's do it properly).