Add DE1-SoC Specific Startup Code
The startup code checks if the load region from the scatter file places the image in the ARM on-chip RAM, and if so will make a backup copy into the bootloader cache region. This is intended to help the first watchdog example such that when the watchdog resets, it relaunches the application image.

Note there are a couple of flaws with this approach:

1. It assumes no code is placed in the DDR memory - ephemeral stuff like stack and heap are fine, but nothing RO/RW
2. If you select "Debug from symbol" in the debug configurations, the debugger temporarily overwrites the image with a breakpoint instruction (which it is supposed to restore later). Unfortunately if this breakpoint is placed after the bootloader "__init_board" copy step, that breakpoint gets copied into the cache.

Ultimately the purpose of this is for a single demo with the watchdog timer, so these restrictions are acceptable. For the debugger issue, the best approach is to select "Debug from entry point", then add "thbreak main" in the "Execute Debugger Commands" box. The thbreak command doesn't modify the image, it uses the DAP which is far more sane anyway.