Add Batch Script for generating U-Boot Image
This is the post-build script useful for generating user application images to be placed on the DE1-SoC SD card to allow operating standalone if so desired. Note that the required binaries are from a number of sources: eadelf is from the armgcc compiler, mkimage is from Quartus SoC EDS, and sed/grep are from git-scm. These are all software components used as part of the ELEC5620 teaching, we just collect the binaries here to avoid issues with different install paths. The binaries here are intended for this use only and not for distribution.