Add Output Cache to FPGA PIO Driver
When using a mode with both input and output signals, it is not possible to directly read back the output register value due to hardware constraints. Previously this limited support for using bit masking on the output.

To combat this, we now keep a cached copy of the previously written output value in the driver context. This allows us to perform R-M-W with the cached value when we can't access the actual register.