Change LT24 Driver to use GPIO Instance
Rather than performing the GPIO bit modification in the LT24 driver, farm this off to a GPIO driver instance. In the actual hardware this is how it is performed - there is an FPGA_PIO controller used for the bit-bashing interface.

This also means that if anyone wishes to use the touchscreen interface bits, they can simply use the same GPIO controller without needing to modify the LT24 driver.