Fix Instruction Cache Issue
Sometimes bare metal images would randomly fail to launch, raising an undefined instruction exception in the bootloader. This seems to be caused by an instruction cache issue.

While we previously disabled all the caching to try and fix this issue, for the Cyclone V this is not enough. We must first enable the caches otherwise disabling fails silently.